I updated the layout of the user configuration layout
The actions connect to partial views to control the background
computations. Then, it redirects to the main config page.

Also, the config options are operational except for the edit user page.
I'm not sure about what should be edited other than user type. The
problem is I have no idea how to access the user type.